- change type-hint for `Config` - change name of an expected file - update documents
