+++ changes/966-koxudaxi.md
@@ -1,0 +1,1 @@
+Mypy plugin support for dataclasses
