+++ tests/mypy/outputs/fail4.txt
@@ -1,0 +1,3 @@
+20: error: Unexpected keyword argument "name" for "AddProject"  [call-arg]
+20: error: Unexpected keyword argument "slug" for "AddProject"  [call-arg]
+20: error: Unexpected keyword argument "description" for "AddProject"  [call-arg]
