+++ tests/mypy/outputs/plugin-fail-strict.txt
@@ -32,4 +32,4 @@
 185: error: Unexpected keyword argument "x" for "AliasGeneratorModel2"  [call-arg]
 186: error: Unexpected keyword argument "z" for "AliasGeneratorModel2"  [call-arg]
 189: error: Name 'Missing' is not defined  [name-defined]
-197: error: Argument "config" to "dataclass" has incompatible type "Dict[<nothing>, <nothing>]"; expected "Optional[Type[BaseConfig]]"  [arg-type]
+197: error: Argument "config" to "dataclass" has incompatible type "Dict[<nothing>, <nothing>]"; expected "Optional[Type[Any]]"  [arg-type]
