--- tests/mypy/outputs/plugin-success.txt
@@ -1,3 +1,0 @@
-121: error: Unexpected keyword argument "name" for "AddProject"  [call-arg]
-121: error: Unexpected keyword argument "slug" for "AddProject"  [call-arg]
-121: error: Unexpected keyword argument "description" for "AddProject"  [call-arg]
