+++ tests/mypy/test_mypy.py
@@ -25,7 +25,8 @@ cases = [
     ('mypy-default.ini', 'fail1.py', 'fail1.txt'),
     ('mypy-default.ini', 'fail2.py', 'fail2.txt'),
     ('mypy-default.ini', 'fail3.py', 'fail3.txt'),
-    ('mypy-default.ini', 'plugin_success.py', 'plugin-success.txt'),
+    ('mypy-default.ini', 'fail4.py', 'fail4.txt'),
+    ('mypy-default.ini', 'plugin_success.py', None),
 ]
 executable_modules = list({fname[:-3] for _, fname, out_fname in cases if out_fname is None})
